new submit button, remove toolbar buttons on mobile, placeholder user icon
